+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 20141002
 * Requirements files
+* License agreement
 
 ### 20141001
 * Added deb installers to Text Editors